project: cleanup: refactor for the network solution
Refactor code to accomodate for the new networking solution:

	* Remove unused code;
	* Format code;
	* Optimize data query's using stored functions.

+++ Common/SensateService.Common.Core/Infrastructure/Repositories/ISensorLinkRepository.cs
@@ -17,6 +17,6 @@ namespace SensateService.Infrastructure.Repositories
 	public interface ISensorLinkRepository
 	{
 		Task<IEnumerable<SensorLink>> GetByUserAsync(SensateUser user, CancellationToken token = default);
-		Task<int> CountAsync(SensateUser user, CancellationToken ct = default);
+		Task<long> CountAsync(SensateUser user, CancellationToken ct = default);
 	}
 }
